sanic_app/requirements.txt
sanic==23.3.0
sanic-ext==23.3.0
PyJWT==2.6.0
